+++ packages/firebase_analytics/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.4
+
+* Bump the minimum flutter version to 1.2.0.
+* Add Template type parameter to `invokeMethod` calls.
+
 ## 2.0.3
 
 * Add resetAnalyticsData method